requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
circuitbreaker==1.4.0
click==8.0.4
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
hf-xet==1.1.1
html5lib==1.1
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langdetect==1.0.9
langfuse==2.60.4
langgraph==0.3.34
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langgraph-supervisor==0.0.21
langsmith==0.3.32
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
moviepy==2.1.2
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oci==2.150.1
oci-cli==3.54.3
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
openpyxl==3.1.5
oracledb==3.1.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250506
pexpect==4.9.0
pi_heif==0.22.0
pikepdf==9.7.0
pillow==10.4.0
platformdirs==4.3.8
proglog==0.1.11
prompt-toolkit==3.0.43
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyOpenSSL==24.1.0
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.5.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.4.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.3
setuptools==78.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
streamlit==1.45.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminaltables==3.1.10
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.34.0
unstructured-inference==0.8.10
unstructured.pytesseract==0.3.15
urllib3==2.2.2
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
